+++ test/edc-lib/exec1by1.cpp
@@ -20,7 +20,7 @@ std::list<SchedJob*> * jobs = nullptr;
 SchedJob * currently_running_job = nullptr;
 uint32_t platform_nb_hosts = 0;
 
-uint8_t batsim_edc_init(const uint8_t * data, uint32_t size, uint8_t flags)
+uint8_t batsim_edc_init(const uint8_t * data, uint32_t size, uint32_t flags)
 {
     format_binary = ((flags & BATSIM_EDC_FORMAT_BINARY) != 0);
     if ((flags & (BATSIM_EDC_FORMAT_BINARY | BATSIM_EDC_FORMAT_JSON)) != flags)
